+++ packages/frontend/app/home/page.tsx
@@ -1,0 +1,5 @@
+import { Home } from "./components/home";
+
+export default function Page() {
+  return <Home />;
+}
